Refactoring of notebooks
Moved the test under modules directory
Removed the added check for VPC and subnet, since it can be optional
